pyproject.toml
[tool.poetry]
name = "isic-2024-skin-cancer-detection-with-3d-tbp"
version = "0.1.0"
description = "Kaggle competition"
authors = ["vip30service123 <80677107+vip30service123@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
catboost = "1.2.5"
dvc = "3.51.2"
hydra-core = "1.3.2"
lightgbm = "4.4.0"
matplotlib = "3.9.1"
mlflow = "2.14.2"
omegaconf = "2.3.0"
opencv-python = "~4.10"
pandas = "2.2.2"
python = "~3.10"
tensorflow = "2.16.2"
tensorflow-io-gcs-filesystem = "0.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
